requirements.txt
# redis
redis==5.0.0
django_redis==5.3.0
celery==5.3.1

# django
Django==4.2.1
drf-yasg==1.21.5
pytest-django==4.5.2
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.2.2

asgiref==3.6.0
packaging==23.1
Pillow==9.5.0
PyJWT==2.6.0
pytest==7.3.1
pytz==2023.3
boto3==1.28.37
psycopg2

# elasticsearch
django-elasticsearch-dsl-drf
django-elasticsearch-dsl
elasticsearch-dsl